Refactor calendar component in HomePage
